pyproject.toml
[tool.poetry]
name = "test-pyro"
version = "0.1.0"
description = ""
authors = ["A-V-tor <avtorca4@gmail.com>"]
readme = "README.md"
packages = [{include = "test_pyro"}]

[tool.poetry.dependencies]
python = "^3.11"
tgcrypto = "^1.2.5"
pyrogram = "2.0.0"
aiohttp = "^3.8.5"
apscheduler = "^3.10.1"
python-dotenv = "^1.0.0"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
